The old handling was to pass the correlation ids through the proxy unchanged, which prevents the proxy (and filters in particular) from making any of their own requests. (If a filter _did_ make its own requests it would have needed its own correlation id, meaning either being able to allocate one that the client couldn't possibly be using, or that the client would end up seeing a correlation id it didn't expect, resulting in exceptions on the client). So in order to allow filters to issue their own requests using the same connection to the broker as is being used for the client requests we need to decouple the upstream correlation ids from the downstream ids.
I haven't yet implemented the `FilterContext` API which will be used to allow filters to make their own requests. Consequently it happens that the correlation ids allocated by the proxy with this new handling will exactly match those from the client (well, the AK client at least, which allocated correlation ids incrementing from 0). So in that respect this commit doesn't actually change anything. Follow-up work will add support for `FilterContext.forwardRequest` and `FilterContext.sendRequest`, the former for forwarding downstream requests from a filter, and the latter for making new ones (where there is no downstream request) and delivering the response to the relevant filter.

To summarise the basic idea we're applying here:

* Record the original correlation id on the `RequestFrame` at the point of decoding. We have to do this for both decoded and opaque frames, so we don't use the RequestData (also the `RequestData`'s correlation id is mutable by filters, so it's defensive programming to not rely on that).
* At the point we serialize the request for sending to the upstream broker, keep in the `CorrelationManager` the information we'll need later, and allocate a new upstream correlation id to use for the request.
* When decoding a response, look up the upstream correlation id in the `CorrelationManager`, to recover the downstream correlation id, and set that on the `ResponseFrame`.

There is a bit of an annoying wrinkle: Kafka is not purely a request/response protocol. A `Produce` with `acks=0` will have no response. For this case we need to avoid filling up the `CorrelationManager` with requests which will have no response (=> OOME). For this reason we need to know, in the RequestEncoder whether ths request we're sending upstream will have a response, which required decoding enough of a `Produce` request to know the `acks`.