Add NetFilter abstraction for pluggable connection logic
NetFilter provides a plugin abstraction for how the connection to a
backend Kafka cluster happens.

It can consume information from earlier handler in the Netty pipeline
including:

* Network peer information from the TCP layer
* SNI information from the TLS handshake
* Client information propagated by HAProxy using the PROXY protocol
* SASL authentication information from the KafkaAuthnHandler

Using this a `NetFilter` implementation can decide whether and which
backend cluster the client should be connected to, and with what
protocol filters.

Because clients may send `ApiVersions` requests before any other request
this means we need to be able to send an `ApiVersions` response from the
proxy without knowing which backend cluster the client will ultimately
connect to. For now we're just using a fixed response read from a
classpath resource, however we should really improve this, but example by
figuring out an "inclusive response" from a set of backend clusters at
runtime.

Note that this means we have decoupled client connection and
authentication from cluster connection. Therefore bad clients
without valid credentials have no effect on any cluster.

Also provided is a `SimpleNetFilter` implementation of the `NetFilter`
interface, which support a single cluster and a fixed list
of protocol filters.

Fixes #68
Fixes #67