Initial implementation of #74: SASL Authentication offload (#88)
Introduce `KafkaAuthnFilter` which should be configured at the front of the
pipeline (after the request decoder) and which intercepts `ApiVersions`,
`SaslHandshake` and `SaslAuthenticate` frames in order that the proxy
authenticates the remote client.

Currently only SASL-PLAIN and SASL-SCRAM-SHA are supported and there is no
easy way of configuring the "user database" (though the test shows how this
can be done). We borrow some of the SASL code from Kafka in the
implementation.

When authentication is successful a Netty user event (`AuthenticationEvent`)
is fired down the pipeline to later handlers. This is not consumed by
any filters in this commit but will be in a later commit.

This commit also introduces the `BareSaslRequest and `BareSaslResponse`.
These are to cope with the history of how Kafka handles SASL.
Before KIP-152 the SASL bytes where sent down the wire without any Kafka
request encapsulation. This doesn't fit nicely with the class hierarchy
we have for `RequestFrame` and `ResponseFrame`, which assume the presence
of Kafka headers. Although `BareSaslRequest` and `BareSaslResponse` are
used here in the `KafkaAuthnFilter`, they're not yet supported by the
`KafkaRequestDecoder` or `KafkaResponseEncoder`, and so the proxy
continues not to support such old clients with this change.