Put per-filter fields in FilterInvokerHandler, superclass of FilterHandler
This reduces the impact of https://bugs.openjdk.org/browse/JDK-8180450
because the ssc is only disturbed when a handler is created for each filter
(i.e. during connection) rather than every time a message is intercepted by
a filter.

Support for API version-aware interception is supported via @ApiVersions.

KrpcFilter becomes a marker interface.

The FilterType enum is introduced to make it easier to iterate and switch over
(in effect) <API, message type> tuples.