Enable configuration of micrometer global registry
This enables configuration of some key micrometer features:
1. common tags to add to all metrics (labels in prometheus-speak)
2. binders (components that can provide common metrics like jvm or system metrics)

and adds a configurable hook so you could apply arbitrary configuration to the
global registry.

You can configure common tags to add to all metrics by adding configuration like
this to the kroxy yaml:

```
micrometer:
  commonTags:
    zone: euc-1a
```

You can configure binders to bind to the global registry by adding configuration
like this to the kroxy yaml:

```
micrometer:
  binders:
    - JvmGcMetrics
    - io.micrometer.core.instrument.binder.jvm.JvmHeapPressureMetrics
```

First we check if there is a class that can be loaded directly using the
string as a type name, otherwise we will try to load it from these packages:

* io.micrometer.core.instrument.binder.jvm
* io.micrometer.core.instrument.binder.system

Note it is assumed that the binder implementations have a no-arg constructor
if you need something more complex you could use the configuration hook.

You can specify a MicrometerConfigurationHook implementation as a way to perform
arbitrary configuration of the global MeterRegistry by adding configuration
like this to the kroxy yaml:

```
micrometer:
  configurationHookClass: your.custom.HookClass
```

This hook will be called after the `binders` and `common tags` are configured