Load Filters reflectively
With this change you can supply the full classname of a Filter and it
will be loaded. The FilterConfig class is inferred from a constructor
of the KrpcFilter implementation

To select the constructor of the KrpcFilter implementation we look for:
1. A constructor annotated with FilterCreator
2. The only constructor if none are annotated (includes the implicit 0-arg
   constructor if you don't define one)

Then, we allow that constructor to contain a single parameter that extends
FilterConfig. We select that subclass as the Filter's config class.

When constructing the Filter we then fill in the config parameter, and can
also optionally fill the ProxyConfig as an argument. If we find any other
parameters of unexpected types we throw an error.

Note that for compatibility we can still refer to the existing builtin
filters by a shortname and this is mapped to the matching filter class.
So kroxy can make it more readable to load some core filters.

Why:
We want to enable FilterAuthors to use their newly minted filters without
having to modify the BuiltinFilterContributor.