Change kroxylicious Future to jdk CompletionStage in KrpcFilterContext
Why:
We are working to extract an filter api that Filter Authers will use to
build custom filters. This currently would include the kroxylicious futures
classes, so adds the extra complexity of extracting that into a module or
making kroxy futures part of the filter-api. With this change we use a JDK
supplied CompletionStage in the KrpcFilterContext