Refactor to introduce generic Contributor and BaseContributor
The BaseContributor is a convenience for creating concrete Contributor
implementations using a typesafe builder to help ensure you are wiring
the right Config subclass to your Filter constructor.

The Contributor and BaseContributor live in a new io.kroxylicious.proxy.service
package as they didn't really fit in an existing package and are
related to loading services.

FilterConfig becomes BaseConfig, so all configuration classes can
extend it.

Why:
We want to re-use the contributor mechanism to load other services
that require some configuration from the Kroxy YAML.