Extract kroxylicious-api module to contain Contributor and Configs
Why:
We are going to re-use the Contributor for other services than filters
so it makes some sense to have a generic kroxylicious-api module rather
than including it in a module called kroxylicious-filter-api.

Note:
I extracted a ProxyConfig interface because the existing class looked
too detailed to me, do all Contributors need to know if ioUring or
frame logging is switched on?