Change performance tests results branch to performance-results
Why:
Performance benchmark results are wiped from the gh-pages branch by the asciidoc
action when commits are pushed to main. The asciidoctor-ghpages-action action
rebases the branch down to a single commit and blows away other changes. So commits
from the performance tests are lost.

By writing to the performance-results branch we will keep historical results but the
nice graphs won't be available in github pages. We can check them out locally to view
them.