Refactor from kroxylicious futures to JDK futures
Wrap the CompletableFuture in our own InternalCompletionStage so that
we can throw an UnsupportedOperationException when the Filter Author
tries to convert it to a CompletableFuture.

Why:
We don't want to maintain our own futures code at this point. The
kroxy implementation has the advantages of a Promise/Future separation
with no blocking methods available to shoot yourself in the feet with.
By preventing the Author working with a CompletableFuture they will be
encouraged to not block in their filter and avoid doing other odd behaviour
like completing the future unexpectedly.

A downside is this introduces another object that needs to be collected.