Refactor from kroxylicious futures to JDK futures
Note we use BrutalFuture a small subclass of CompletableFuture to
override the standard completion methods and `get` method to try and
control when it can block and prevent clients from completing the
future from their filters.

Why:
We don't want to maintain our own futures code at this point. The
kroxy implementation has the advantages of a Promise/Future separation
with no blocking methods available to shoot yourself in the feet with.