Set language level to 17, 19 for integration tests
Why:
The split language levels across test and main sources has confounded
IntelliJ. It hangs together okay in an experimental maven import mode
but recently we found an issue where test resources were not on the
classpath when running some new unit tests. Moving to a setup with the
same language level for test/main per maven module works more smoothly.

We also want to target a LTS java version as a minimum runtime version.

The integrationtests use a java 18+ feature to configure a custom
InetAddressResolver for resolving some test hosts, so we are using
a higher language level there.