Enable krpc generator mojo to configure which source roots it adds to
Why:
We had a problem where javadoc generation was failing because it was processing
generated code unexpectedly. We intended the source to be test only and were
surprised it was included in the javadoc run. We used a build helper plugin to
add the path to the test source roots, but the KrpcGeneratorMojo also added it
to the compile source roots and so it was picked up when generating javadoc.

Now the KrpcGeneratorMojo behaviour is configurable and you can declare which
source roots should have the output dir added to them like:
```
<addToProjectSourceRoots>testCompile</addProjectSourceRoots>
```
or
```
<addToProjectSourceRoots>compile</addProjectSourceRoots>
```
So we can add generated sources to the test source root only and avoid having it
selected for javadoc generation.

Relates to #229