multitenant: configure krpc plugin to add it's output dir to test source roots
Remove the build helper that was setting up the test source root

Why:
The krpc generator plugin adds it's output dir to the compile source roots of
the project by default. Theses sources are selected by the javadoc plugin when
generating javadoc. This caused javadoc generation to fail as the generated classes
depended on some test scoped dependencies which weren't made available to the
javadoc generator.

Closes #229