Wait for services in performance action
Why:
We saw a failure where kroxylicious was never started due to trying to
reference a non-existent JAR. Because this was backgrounded it didn't
cause the action to fail and the kafka producer performance script will
retry forever, so the action eventually timed out after 6 hours. We want
to fail faster to conserve action time and expose the issue.