Add kroxylicious-test-tools module containing low-level KafkaClient
KafkaClient is a test client oriented around sending a single Kafka RPC to a Kafka
broker, receiving a response and then closing its channel.

I chose to copy the relevant codec classes from kroxy rather than attempt to reuse them so that:
1. the mock code is independant of kroxy code, so we know which variable is changing when testing
2. the mock doesn't need some proxy code like correlation id rewriting, we can focus the
code better. We don't want this test utility to get in the way of changes to kroxy which is still
changing rapidly.
3. prevents a circular dependency or the hassle of extracting and maintaining a codec module