Add ApiMessageSampleGenerator to kroxylicious-test-tools
Why:
To support testing that kroxylicious can decode and modify every kafka
RPC at every supported apiVersion it is nice to have a valid sample of
each RPC/apiVersion. The generator uses reflection to generate an example
of each ApiMessage at each version, filling in all fields that are part
of the schema for that version. The data is generated randomly using a
Random with a fixed seed, so the generation should be deterministic on
a given JVM.