Add mock kafka broker to kroxylicious-test-tools
MockServer is a netty server that mocks out the kafka broker. You can configure
it with one response at a time. It stores references to messages it receives for
later verification, this can be cleared between test runs.

MockServer has a stripped down copy of the Kroxylicious codec internals for
encoding and decoding requests/responses. I explored using Kroxylicious itself
as a mock server but since the code is proxy-oriented it would by pushing against
the grain to write filters that short circuit the proxy. Or you could make Kroxy
support a 'mocking mode' where there is no proxied server, again this would make
the Kroxy code less focused on proxying and could be a barrier to developing proxy
features.