Add integration test for simple proxy and manipulation of all RPCs
Introduces a FixedClientIdFilter and TestFilterContributor to integrationtests
so we have a filter available that decodes all requests/responses and modifies
the header of all requests.

Why:
We want to validate that Kroxy in a very simple setup can decode all
Kafka RPCS, at all supported apiVErsions, apply some simple manipulation,
re-encode the request, proxy it to the physical broker and forward the
response back to the client.

These tests use a mock kafka broker with a configurable response and
a test client to send a single RPC at a time to kroxy, and then verify
the message received at the mock is what we expect and that the response
recieved by the client is what we expect.