Enable filters to intercept/mutate SASL requests when proxy is not handling SASL
Why:
Until now the proxy was only considering if the proxy was handling SASL
offload when deciding if it should decode SASL requests. If no mechanisms
are configured (currently hardcoded to no mechanisms) then SASL requests
are never decoded even if you have a SaslHandshakeRequestFilter or
SaslAuthenticateRequestFilter defined.

This changes two behaviours:
* On the first hit, before we have a delegate set on the SaslDecodePredicate
we always decode all RPCs. This means the SASL requests will be available for
filtering on the first hit.
* When the delegate is set, we decode if the delegate says we should decode
or if kroxylicious is handling SASL and it's a SASL request.

Closes #261