Add produce filter that can validate records meet some criteria
Some organisations might want stronger guards against invalid records being produced
to Kafka. The existing schema support is client side and there is no protection against
some client coming along and pushing unexpected bytes to your topic, which could require
manual intervention at the consumer if it cannot progress past the unexpected record.

If the filter finds any invalid records in a Produce request:

1. If the filter finds that all topic-partition data in a request has any invalid record
   in them, it will generate a response without proxying.
2. If the request has a transactionalId, it generates a response marking all topic-partitions
   as failed with error and responds to the client without proxying.
3. If the user has specified that they do not want to forward partial produce data then all
   topic-partitions in the request are marked as failures and the Filter responds to the client
   without proxying
4. If the user has specified that they do want to forward partial produce data then all
   valid topic-partitions are sent to the proxied broker. When the response from the broker
   is intercepted, all failed topic-partitions are augmented in with corresponding error codes.

This feature is built on the assumption that one topic should have one validator. We use
Rules to determine which validation to apply to a given topic. By iterating through the
rules in the order they are declared, checking if a topic's name matches each rule, we locate
the first matching rule. If no rule is matched this way we can configure a default validation
rule to be applied.

For example, this configuration has a rule for topic "one" and topic "two":

```
- type: ProduceValidator
  config:
    defaultRule:
      valueRule:
        allowNulls: true
        allowEmpty: false
    rules:
    - topicNames:
      - one
      valueRule:
        syntacticallyCorrectJson:
          validateObjectKeysUnique: true
    - topicNames:
      - two
      valueRule:
        allowNulls: false
        allowEmpty: false
```

The first validation implemented is JSON syntax validation, checking if the value is well-formed JSON.
The user can optionally enable it to validate the uniqueness of keys in JSON objects.

This is built with flexibility in the YAML configuration and implementation in mind, so that further
validations can be plugged in later. For example it would be great to integrate with Schema
registries to check that the data is encoded using a schema.