Refactor: invoker on[Request|Response] is responsible for checking if it needs to act
Why:
It felt like an awkward API asking the invoker if it was okay to invoke and then only
calling on[Request|Response] if it can safely handle the message. Instead we can wrap
with an invoker that can handle any message type and do nothing but forward the message if it has no logic
for this message.

This adds a cost, we have to construct a DefaultFilterContext per message per filter in the chain.
This could be prevented if we move FilterInvoker from the filter api into the kroxylicious
module and make it responsible for building the filter context when required. Currently it
would be a bit ugly to do this as filter-api has no knowledge of the frame.