Refactor FilterInvoker implementation for speed
Instead of switching on ApiKey and checking what type the filter is with
instanceof, this implementation creates an array of request FilterInvokers
and an array of response FilterInvokers. It stores an invoker for each
implemented interface in the array, using the ApiKey.id as index. Otherwise
it stores a handleNothing invoker.

Also adds a JMH performance test for comparing invoker performance in future