Add javadoc and comment around SafeInvoker
This was used to enable the FilterHandler to have a better API
when invoking, changing from:
```
if (invoker.shouldHandle(x,y,z) {
  invoker.onRequest(x,y,z);
} else {
  ctx.forwardRequest
}
```
to
```
invoker.onRequest(x, y, z);
```
Putting the behaviour of unhandled decoded requests into the Invokers
responsibilities.

We chose this composed approach rather than duplicating this logic or using
some inheritance pattern, but they may be easier approaches to understand
if this proves too arcane.