Add apiVersion to specific message filter onXyz[Request|Response] methods
Why:
Currently we do not have a way to access the apiVersion for Responses.
The apiVersion is in the request headers but so far we have not made
it available to response filters. This is a problem when we want to
construct response data in the proxy, we need to know which fields can
be serialized at the apiVersion used by the client. Also we have started
adding code which is trying to sniff the version by checking for non-null
fields, sometimes it might be cleaner to switch on the apiVersion instead.

I opted to add it to the signatures for requests and responses for symmetry
despite apiVersion also being present in the request header data.