Introduce KroxyTester to simplify testing Filters and the proxy
This expands the responsibilities of the kroxylicious-test-tools module
to become a library Filter Authors can use to integration test their
Filter. It now offers:

1. A fluent API to build kroxylicious configuration YAML
2. Conveniences for configuring Kroxylicious for the simple case
where one kroxy proxies one broker and wiring in the standard ApiVersions
and BrokerAddressFilter.
3. A KroxyTester interface for offering convenient access to clients
for Kroxy so users do not have to wire in bootstrap address and serializers
everywhere.
4. Implementations of KroxyTester for testing against an external cluster
(typically wired through the kroxy junit extension) and testing against a
mock kafka broker.

Why:
We want to enable Filter Authors to test their filters against
Kroxylicious and a proxied Kafka cluster. KroxyTester will remove
some of the boilerplate required to create and configure the proxy
and backing server, then make it easier to create clients and interact
with the cluster.