Add scripts and manifests to support running kroxylicious on minikube
These scripts enable developers to run kroxylicious against a Strimzi managed
Kafka cluster on minikube. This should help us test different topologies in a
reproducible way.

1. scripts/deploy-image.sh

This script builds and deploys the Docker image to quay.io. It is required for
a user to have a quay.io account and a public `kroxylicious` repository. This
is a way to enable minikube to pull development images easily.

2. scripts/run-with-strimzi.sh

This script invokes deploy-image.sh to build and push a development image, then
it starts minikube and installs a 3-node Kafka cluster using Strimzi. Then it
installs your kroxylicious image, configured to proxy the Kafka cluster.