Handle kafka 3.5.0 spec changes
1. ConsumerGroupHeartbeatResponse added the first nullable Struct, the
generator had some code that detected that nullableVersions was set
for a field but we have not explicitly marked it as nullable.

2. A new field was added to MessageSpec; "latestVersionUnstable" to mark
it as an API under development.

3. `ApiMessageType#highestSupportedVersion()` has been removed/replaced
with `ApiMessageType$highestSupportedVersion(boolean enableUnstableLastVersion)`
I've made Kroxylicious support the unstable last version as it should be the
decision of the backing broker whether to accept unstable versions, the
proxy should handle them.

4. MultiTenantFilter had to cope with field names changing on
`AddPartitionsToTxnRequestData` and `AddPartitionsToTxnResponseData`.
Still TODO is handle v4+ of this API

Also added assertions to the existing code generation tests so we are
verifying something about their output, not just that the generator
runs without error.