FilterContributor contribute List<KrpcFilter>
In some cases it may be preferable if we can contribute multiple
filters for a single Filter definition in the config.

For example for multi-tenancy we need to use per-MessageType filters
to intercept specific RPCs. But we also will want to map all clientIds
in request headers, which makes more sense as a RequestFilter that
intercepts all request. These types of filters cannot be mixed in a
single implementing class currently, so the Users would have to add
two filter definitions to their YAML.

It would be better if they could declare a single type and have the
required filters installed into the chain in the right order. The
configuration is simplified and better for evolution too. It allows
us to decompose a single filter into multiple, or recombine them
without the user configuration having to change.

Closes #348