Add CompositeFilter
Filter Authors may now implement CompositeFilter if they wish a single
filter definition in YAML to contribute multiple filters to the filter
chain for a channel. CompositeFilter is not allowed to be combined with
RequestFilter, ResponseFilter or any specific message Filter interfaces.

Why:
Filter Authors may wish to implement a single package of behaviours
as multiple discrete KrpcFilter implementations. This can aid readability
by keeping classes focused. We also have restrictions about what interfaces
are allowed to be implemented on a single class. RequestFilter and ResponseFilter
are incompatible with the specific message Interfaces as it would be ambiguous
which method should be called for any message handled. CompositeFilter enables
us to offer them as a bundle, simplifying configuration and enabling us to
evolve the filter (changing the filters used to implement it's behaviour).