Throw on port collision with admin http port
If the admin port is enabled and any virtual cluster has declared
that it wants to bind to that port (either with a shared or exclusive
port), then PortConflictDetector throws an IllegalStateException.

If either the admin endpoint or the virtual cluster is binding to
the <any> interface (ie 0.0.0.0), we check if their ports collide.

If both the admin endpoint and the virtual cluster are binding to
the same specific interface (not 0.0.0.0), we check if their ports
collide.

Why:
We saw issues where we bound to bootstrap localhost:9192 with
the prometheus port active on the default port, 9193. So when
we attempted to bind the first broker to 9193, it failed with a
port conflict. We can check this earlier by ensuring that no
cluster will bind to the admin http host/port if the admin http
port is activated.