Change default admin http port from 9193 to 9190
Why:
Since the addition of the port-per-broker cluster binding mechanism
the default metrics port of 9193 has been colliding with the port
range used for example config and documentation, with the bootstrap
on localhost:9192 and therefore broker-0 will bind on 9193. It might
be better to change the default metrics port so people can continue
using 9192 for proxying.