Fix buffer leak on short circuit response
A leak was introduced when we added the ability to `forwardResponse`
while handling a Request. If the filter allocates a buffer using
KrpcFilterContext#createByteBufferOutputStream, then the buffer is
added to the DecodedFrame associated with the context. Then it is assumed
that this frame will be read or written to the netty channel, so that
netty can call release on it, which releases the buffers on the frame.
In the short-circuit response case we create a new frame and the buffers
on the old frame are left to be garbage collected.

With this solution we transfer the buffers to the new frame that is
handed to netty.