Hack in handling connections on broker ports before metadata available
Initially we bind the broker ports for port-per-broker to map to the upstream
bootstrap address. If we receive connections on these broker ports then we
discard the first message from the client and request metadata. Once we have
reconciled the metadata response we close the channel. Now subsequent connections
should have an up to data binding.

Note: what if the upstream bootstrap address happens to be the same as a broker
address, we want to make sure we don't keep doing this behaviour.