Fix #412: remove $(portNumber) replacement pattern in the PortPerBroker scheme. (#428)
* Fix #412: Removes the $(portNumber) pattern from brokerAddressPattern parameter PortPerBroker scheme.

Why: it was superfluous configuration. The only permitted value was $(portNumber). It was substituted for the
port number allocated to the broker from the port range. Removal gives us simpler configuration without loss of function.

Also in the SniRouting case, we now validate that there is no port specified on the same parameter. SniRouting always
uses the port number assigned to the virtual cluster's bootstrap. If a port number was specified, it actually lead to a
failure later (client would choke on a badly formed addressed host:xxxx:xxxx.

Signed-off-by: kwall <kwall@apache.org>
Co-authored-by: Robert Young <robertyoungnz@gmail.com>