Obtain upstream ApiVersions when proxy is not SASL offloading
If Kroxylicious is not handling SASL offload (currently it never handles SASL offload) then
we initiate NetFilter#selectServer if we receive an ApiVersions request and buffer the message.
After the connect is successful we send the buffered message to the backend. We also install the
ApiVersionsFilter if we are not handling SASL offload so that we are intersecting the Kroxylicious
supported versions with the upstream versions without further configuration.

Why:
Currently Kroxylicious responds with a hardcoded ApiVersions response. This is intentional, to
support the proxy handling SASL authentication before we make a connection to the upstream
broker, protecting the broker from unauthorized requests and enabling NetFilter to use the
authorized id as a criteria for backend selection. The response is hardcoded to the ApiVersions
values for a kafka 3.2 broker.

This means that if we are proxying a broker older than v3.2, Kroxy will be declaring it supports
api versions that it doesn't know about, causing errors when you try to use a newer client that
supports those newer API versions. Conversely if the backing broker is newer and supports new
api versions, the client would be told that the broker only supports v3.2 APIs and the client
can't take advantage of new APIs.

Addresses #441