Add test for suspicious api versions behaviour
The ApiVersions filter currently passes through broker versions for ApiKeys
that Kroxylicious does not recognise. This seems a little inconsistent that
we limit all the known APIs to the versions supported by Kroxylicious but then
allow brand new APIs to be exposed to the client. The proxy would also fail
to read those messages as we attempt to convert from short to ApiKeys in the
KafkaRequestDecoder, which would throw. So another inconsistency, Kroxylicious
would be declaring it can handle an ApiKey that it cannot.