Enable Response filters to intercept out-of-band responses
Why:
There was an inconsistency in that out-of-band requests could be intercepted
by upstream filters, but the Cluster's response was not intercepted. The javadoc
for sendRequest said `The response will pass through upstream filters prior to
the handler being invoked`, which can be read as saying they should be invoked.

Closes #445