Use discovery endpoints/SNI matching to allow existing clients to reconnect using broker addresses they already know / and support the use-case where Kroxylicious is scaled. (#447)
Fix #416, #417 

In the PortPerBroker case, the ClusterNetworkAddressConfigProvider now announces a `discoveryAddressMap`.  If the provider provides these, those endpoints will be bound with a special kind of restricted broker binding.

In the SNIRouting case, in  EndpointRegistry#resolve, if the incoming SNI name doesn't match an existing binding it now falls back on a match with `brokerAddressPattern`. If it matches the  EndpointRegistry#resolve produces a restricted broker binding.

Kroxylicious handler layer understand how to handle restricted bindings. If one is encountered, it inserts an EagerMetadataLearner into the filter chain.  This filter's job is to hijack the connection after SASL exchange, and dispatch an out-of-band Metadata request.  Once the Metadata response is received, it closes the connection.  This relies on the fact that BrokerAddressFilter (upstream of it) will have reconciled metadata.

Signed-off-by: kwall <kwall@apache.org>
Co-authored-by: Robert Young <robeyoun@redhat.com>