Support configuring multiple responses in MockServer
Why:
The OutOfBandRequestIT used an unsupported pattern to forward a response
from within a sendRequest callback using a reference to the request context.
By enabling the mocks server to have a separate response configured per api
key we can use a supported pattern, storing the result of the send request
in a member field to be augmented into a response later.