Do not expose unknown APIs to the client
Why:
If the broker is ahead of Kroxylicious and is offering a new RPC, but Kroxylicious
does not know how to handle it we do not want to tell the client that we support
that RPC. If the client sent that RPC to us we would fail to decode the request with
an exception. So we remove unknown ApiKeys from the ApiVersionsResponse when intersecting
with Kroxylicious's supported versions.

Co-authored-by: Sam Barker <sam@quadrocket.co.uk>