Toggle autoread on the channel when deferring
Why:
We do not want to allow arbitrary quantities of requests to
pile up in memory while we are doing some deferred work.
This would be particularly dangerous with zero-ack Produce
Requests where the client does not expect a response. It
could continue sending those until OutOfMemory.

Expert advice is stopping autoread is going to be a performance
killer. With the branch as it is, if we defer then that will
stop the channel doing Filter work until it completes or times out.
Not very efficient if we have multiple batched messages that could
have async work done on them in parallel. So performance is a
worry.