Move CLI entrypoint to new kroxylicious-app module
Why:
When you start up the jar currently emitted by the kroxylicious module with
the dist profile, nothing is logged. We currently do not have an slf4j implementation
supplied. This is nice if users wanted to embed Kroxylicious. But we usually want
logging if we are running kroxylicious in a container etc. So this commit introduces
a module which configures default logging with sl4fj and provides the picocli integration
and logs the banner/versions.

The default log level is WARN except for some specific loggers which are enabled by user
configuration. Users could set the `log4j2.configurationFile` system prop to override the
classpath configuration, or set the env var `KROXYLICIOUS_ROOT_LOG_LEVEL` to change the
default log level (eg `KROXYLICIOUS_ROOT_LOG_LEVEL=DEBUG java -jar ...`).