Add integration test covering Short-circuit problem
Currently if we have proxied a message to the broker we could then
move onto the next message and short-circuit respond. The client will
get the response LIFO, which breaks the contract that they expect
responses back in the order they sent them.

To test this I've widened the abilities of the test KafkaClient so
that it can write and flush multiple messages, await a matching number
of responses and the complete the future. To achieve this was a little
awkward because the Kroxylicious state machine cannot handle reading
multiple messages while processing the first messages. If throws in the
Frontend Handler because it's in the CONNECTING state. So I've made the
client send an ApiVersions request first, await the response and then
write the remaining messages.